eval/requirements.txt
ragas==0.1.21
datasets
pandas
langchain-ollama
langchain-community
httpx
python-dotenv
elasticsearch==8.12.0
sentence-transformers
